api/requirement.txt
annotated-types==0.6.0
anyio==3.7.1
azure-storage-blob==12.19.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.105.0
h11==0.14.0
idna==3.6
pillow==10.2.0
pydantic==2.5.2
pydantic_core==2.14.5
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
qrcode==7.4.2
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.24.0.post1
pillow==10.2.0